Separate context_function, stack_allocator and simple_stack_allocator
